requirements.txt
torch
torchvision
transformers
sentence-transformers
pycocotools
tqdm
numpy
Pillow
matplotlib
